mediatek: add recipe for DIR-X3260
The DIR-X3260 is built on the same SDK as the COVR-X1860 found in
ramips. There is a simple metadata required at the footer of the image
and dlink-sge-image is required to encrypt it to the correct format.

Unlike the COVR however, the COVR metadata must also reside in the
recovery image, which the recovery web server will check. Another
difference is that it is checking that the size of the firmware is
larger than 16 MiB, and the metadata must be placed after 16MiB.

As of writing this commit, a default dist build within OpenWrt falls
below 10MiB.

Thus, we move the append to recovery (which will be fallen through for
factory image) and ensure that the image is at least 16MiB via pad-to.

Signed-off-by: Han Sol Jin <jhs@jinhansol.com>