mediatek: split firmware partitions into kernel and UBI for DIR-X3260
This adds proper UBI support for the DIR-X3260.

This device has backup partitions for the firmware. The stock U-Boot
would only ever boot the kernel from the first partition (main). Before
booting, the main and backup partitions are compared (hashes of kernel
and rootfs), and, if both partitions look valid but have different
hashes, the backup one will be overwritten with the data from the main
partition (usually after an upgrade of OEM software to a newer
version), then a reboot is performed, expecting both hashes to be
identical upon re-checking, thus booting from the main partition. So
the backup would only ever be used if the main partition fails the hash
verification upon boot.

When it comes to using the flash space of the backup partition, we can
split it into a kernel with fake-rootfs, all padded to 8 MiB, and the
remaining space for fwconcat1, that would make a UBI partition with the
UBI partition from the main image flash space, using mtd-concat.

Thus, when flashing OpenWrt via recovery, both partitions will be
erased, OpenWrt flashed to the main flash area, upon reboot the image
will be duplicated to the backup, making them identical after another
reboot, so the image gets booted from the main partition without
restoring anything.

When flashing OpenWrt via OEM web updater, only the main partition will
be written, but (thanks to fakeroot) be considered valid, so it will be
copied to overwrite backup as well.

Co-authored-by: Sebastian Schaper <openwrt@sebastianschaper.net>
Signed-off-by: Han Sol Jin <jhs@jinhansol.com>